Add PollReadable method to Channel
Returns True if there is data to be read from the Channel.

Raises DataAvailable if data is available; The ExtendedStream parameter will be True if the data that is available is available on the StdErr stream (ID=1)

Raises the Error event if no data was available due to an error. The Reasons parameter is a bitmask of one or more LIBSSH2_POLLFD_* constants defining the error(s) that occurred.