Tweaks to Channel.Poll and friends
Channel.Poll() now uses Channel.PollReadable() and Channel.PollWriteable() to poll all streams. Returns True if action on the channel is possible. Check Channel.LastError for a bitmask of LIBSSH2_POLLFD_* constants describing which directions are ready. PollReadable() and PollWriteable() behave the same except only one direction is polled.

If PollReadable() or PollWriteable() detect an error condition the Error(Int32) event is raised, unless the error condition is "channel closed" in which case the Disconnected() event is raised. Data may still be readable from the channel after disconnect.

SSH.Poll(Extends SSHStream) now makes use of the same API as Channel.Poll(). If the SSHStream represents an SFTPStream then the SFTPSession which owns the stream will be polled; if more than one stream is open then Poll() will return True if *any* of the streams are ready.