Make separate SSH.Put() helper methods for SCP and SFTP
Put(Session=Nil, URL, Overwrite=False) is now only for SFTP.

Put(Session=Nil, URL, Length, Overwrite=False) is for either SCP or SFTP; in SFTP the Length parameter is ignored but is no longer an optional parameter.

Making Length non-optional in one and absent in the other resolves the type error when using certain combinations of optional parameters.